feat: use builder pattern for main.rs
